Quiz Functionaltiy
Base code, changes are needed. Full quiz is operational however no CSS and limited bootstrap is applied.

Some variables are applied to the API pull link which can be modified via user input later on. The quiz is dynamic, and shuffles the answer position around as to not make the quiz too easy. There is a developer "tell me the answer" tool, which is for use in development only. This prints to the console. I think this is a good base for our main javascript file.